third_party/WebKit/LayoutTests/inspector/documentation/documentation-catalog-expected.txt

This test verifies URL created by WebInspector.DocumentationCatalog
Property max:
    0: https://docs.webplatform.org/w/api.php?action=query&titles=javascript/Math/max&prop=revisions&rvprop=timestamp|content&format=json
Property insertBefore:
    0: https://docs.webplatform.org/w/api.php?action=query&titles=dom/Node/insertBefore&prop=revisions&rvprop=timestamp|content&format=json
Property eval:
    0: https://docs.webplatform.org/w/api.php?action=query&titles=javascript/eval&prop=revisions&rvprop=timestamp|content&format=json
Property stringify:
    0: https://docs.webplatform.org/w/api.php?action=query&titles=javascript/JSON/stringify&prop=revisions&rvprop=timestamp|content&format=json
Property isArray:
    0: https://docs.webplatform.org/w/api.php?action=query&titles=javascript/Array/isArray&prop=revisions&rvprop=timestamp|content&format=json
Property now:
    0: https://docs.webplatform.org/w/api.php?action=query&titles=javascript/Date/now&prop=revisions&rvprop=timestamp|content&format=json
Property fromCharCode:
    0: https://docs.webplatform.org/w/api.php?action=query&titles=javascript/String/fromCharCode&prop=revisions&rvprop=timestamp|content&format=json
Property keys:
    0: https://docs.webplatform.org/w/api.php?action=query&titles=javascript/Object/keys&prop=revisions&rvprop=timestamp|content&format=json
Property notExist:
    null
Property indexOf:
    0: https://docs.webplatform.org/w/api.php?action=query&titles=javascript/Array/indexOf&prop=revisions&rvprop=timestamp|content&format=json
    1: https://docs.webplatform.org/w/api.php?action=query&titles=javascript/String/indexOf&prop=revisions&rvprop=timestamp|content&format=json
Property setDate:
    0: https://docs.webplatform.org/w/api.php?action=query&titles=javascript/Date/setDate&prop=revisions&rvprop=timestamp|content&format=json
Property parse:
    0: https://docs.webplatform.org/w/api.php?action=query&titles=javascript/Date/parse&prop=revisions&rvprop=timestamp|content&format=json
    1: https://docs.webplatform.org/w/api.php?action=query&titles=javascript/JSON/parse&prop=revisions&rvprop=timestamp|content&format=json
